Resolvendo bugs no backend e adicionando função de autenticação de usuário

+++ backend/public/api/deletar-reserva/index.php
@@ -1,36 +1,35 @@
 <?php
-    require "../../../vendor/autoload.php";
-    use App\Utils\Response;
-    use App\Utils\Reserves;
-    use App\Utils\Auth;
+    require __DIR__ . "/../../../config/cors.php";
+    require __DIR__ . "/../../../vendor/autoload.php";
+    use App\Backend\Response;
+    use App\Backend\Reserves;
+    use App\Backend\Auth;
 
-    $headers = getallheaders();
-    if (!isset($headers["Authorization"])) {
-        echo Response::json(403, "Autorização não concedida!", ["success" => true]);
-        exit;
-    }
+    if ($_SERVER["REQUEST_METHOD"] == "DELETE") {
+        $user = Auth::verify_token();
+        if (!$user) {
+            echo Response::json(403, "Token inválido", ["success" => false]);
+            exit;
+        }
 
-    $token = str_replace($headers["Authorization"], "Bearer ", "");
-    $user = Auth::verify_token($token);
-    if (!$user) {
-        echo Response::json(403, "Token inválido", ["success" => false]);
-        exit;
-    }
+        if (!isset($_GET["id"])) {
+            echo Response::json(403, "ID não informado", ["success" => false]);
+            exit;
+        } else {
+            $id = htmlspecialchars($_GET["id"]);
+        }
 
-    if (!isset($_GET["id"])) {
-        echo Response::json(403, "ID não informado", ["success" => false]);
+        $reserves = new Reserves();
+        $reserve_delete = $reserves::delete_reserve($id);
+        if ($reserve_delete) {
+            echo Response::json(200, "Reserva eliminada com sucesso!", ["success" => true]);
+        } else {
+            echo Response::json(200, "Não foi possível eliminar a reserva!", ["success" => false]);
+        }
+
         exit;
     } else {
-        $id = htmlspecialchars($_GET["id"]);
-    }
-
-    $reserves = new Reserves();
-    $reserve_delete = $reserves::delete_reserve($id);
-    if ($reserve_delete) {
-        echo Response::json(200, "Reserva eliminada com sucesso!", ["success" => true]);
-        exit;
-    } else {
-        echo Response::json(200, "Não foi possível eliminar a reserva!", ["success" => false]);
+        echo Response::json(405, "Método da requesição inválido.", ["success" => false]);
         exit;
     }
 ?>
